+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/filter/DoubleFilterMaker.java
@@ -12,6 +12,7 @@ import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
@@ -70,7 +71,7 @@ public class DoubleFilterMaker implements FilterMaker {
     }
 
     @Component
-    @Scope("prototype")
+    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
     public static class DoubleFilter implements Filter, Bean {
 
         private static final long serialVersionUID = -3161256701795506170L;