+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/filter/RegexFilterMaker.java
@@ -14,6 +14,7 @@ import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
@@ -73,7 +74,7 @@ public class RegexFilterMaker implements FilterMaker {
     }
 
     @Component
-    @Scope("prototype")
+    @Scope(ConfigurableBeanFactory.SCOPE_PROTOTYPE)
     public static class RegexFilter implements Filter, Bean {
 
         private static final long serialVersionUID = 970186550656361331L;